Crate mdbook_preprocessor

Library to assist implementing an mdbook preprocessor.

This library is used to implement a preprocessor for mdBook. See the linked chapter for more information on how to implement a preprocessor.

Modules

book

A tree structure representing a book.

config

Mdbook’s configuration system.

errors

The error types used in mdbook.

Structs

PreprocessorContext

Extra information for a Preprocessor to give them more context when processing a book.

Constants

MDBOOK_VERSION

The current version of mdbook.

Traits

Preprocessor

An operation which is run immediately after loading a book into memory and before it gets rendered.

Functions

parse_input

Parses the input given to a preprocessor.