mdbook_preprocessor::book

Struct Book

pub struct Book {
    pub items: Vec<BookItem>,
}

A tree structure representing a book.

A book is just a collection of BookItems which are accessible by either iterating (immutably) over the book with iter(), or recursively applying a closure to each item to mutate the chapters, using for_each_mut().

Fields
items: Vec<BookItem>

The items in this book.

Implementations

impl Book

pub fn new() -> Book

Create an empty book.

pub fn new_with_items(items: Vec<BookItem>) -> Book

Creates a new book with the given items.

pub fn iter(&self) -> BookItems<'_>

Get a depth-first iterator over the items in the book.

pub fn chapters(&self) -> impl Iterator<Item = &Chapter>

A depth-first iterator over each Chapter, skipping draft chapters.

pub fn for_each_mut<F>(&mut self, func: F)
where F: FnMut(&mut BookItem),

Recursively apply a closure to each item in the book, allowing you to mutate them.

Note

Unlike the iter() method, this requires a closure instead of returning an iterator. This is because using iterators can possibly allow you to have iterator invalidation errors.

pub fn for_each_chapter_mut<F>(&mut self, func: F)
where F: FnMut(&mut Chapter),

Recursively apply a closure to each non-draft chapter in the book, allowing you to mutate them.

pub fn push_item<I>(&mut self, item: I) -> &mut Book
where I: Into<BookItem>,

Append a BookItem to the Book.